requirements.txt
pefile
pycurl
tqdm
google-cloud-storage
sphinx
pytest
pyqt6
pyinstaller
PySide6
jinja2
vcf_parser
WGSE-NG-3rd-party